Patch: Fixed signed/unsigned comparison (Patch by Elizabeth Oldham)
git-svn-id: https://qtpfsgui.svn.sourceforge.net/svnroot/qtpfsgui/trunk@771 479317e0-7f23-0410-9f0e-fdaa22894045

+++ src/TonemappingOperators/mantiuk08/display_function.cpp
@@ -97,7 +97,7 @@ float DisplayFunctionGGBA::display( float pix )
 
 // ========== LUT Display Function ==============
 
-static const size_t max_lut_size = 4096;
+static const int max_lut_size = 4096;
 
 DisplayFunctionLUT::DisplayFunctionLUT( const char *file_name ) : pix_lut( NULL ), L_lut( NULL )
 {
